Rename src folder to lib
This enables existing imports from lib to continue working
