+++ .gitignore
@@ -2,4 +2,3 @@
 html-report
 node_modules
 tmp
-/lib
