+++ bin/remap-istanbul.js
@@ -2,9 +2,9 @@
 
 const Collector = require('istanbul/lib/collector');
 const MemoryStore = require('istanbul/lib/store/memory');
-const loadCoverage = require('../src/loadCoverage');
-const remap = require('../src/remap');
-const writeReport = require('../src/writeReport');
+const loadCoverage = require('../lib/loadCoverage');
+const remap = require('../lib/remap');
+const writeReport = require('../lib/writeReport');
 
 /**
  * Helper function that reads from standard in and resolves a Promise with the
