+++ lib/CoverageTransformer.js
