+++ lib/SparseCoverageCollector.js
