+++ lib/checkThreshold.js
