+++ lib/getMapping.js
