+++ lib/gruntRemapIstanbul.js
