+++ lib/gulpRemapIstanbul.js
