+++ lib/intern-reporters/JsonCoverage.js
