+++ lib/loadCoverage.js
