+++ lib/remapBranch.js
