+++ lib/remapFunction.js
