+++ lib/writeReport.js
